requirements.txt
streamlit==1.45.1
google-generativeai
pinecone-client==2.2.1
pydantic==2.11.4
requests
httpx
numpy
pyodbc
python-dotenv
langchain-community
langchain-google-genai
fastapi
uvicorn